+++ internal/kafkamsgs/headers.go
@@ -1,0 +1,14 @@
+package kafkamsgs
+
+const (
+	HeaderKeyVersion   string = "Version"
+	HeaderKeyEventName        = "Event Name"
+)
+
+const (
+	EventNameAzureADGroupCreated string = "azure_ad_group_created"
+)
+
+const (
+	VersionAzureADGroupCreated string = "1"
+)
